+++ src/Symfony/Bridge/Twig/Tests/Extension/TranslationExtensionTest.php
@@ -114,6 +114,9 @@ class TranslationExtensionTest extends TestCase
             ['{% set vars = { \'%name%\': \'Symfony\' } %}{{ hello|trans(vars) }}', 'Hello Symfony', ['hello' => 'Hello %name%']],
             ['{{ "Hello"|trans({}, "messages", "fr") }}', 'Hello'],
 
+            ['{{ "bar"|trans }}', 'bar'],
+            ['{{ "bar"|trans(default="fallback") }}', 'fallback'],
+
             // trans filter with count
             ['{{ "{0} There is no apples|{1} There is one apple|]1,Inf] There is %count% apples"|trans(count=count) }}', 'There is 5 apples', ['count' => 5]],
             ['{{ text|trans(count=5, arguments={\'%name%\': \'Symfony\'}) }}', 'There is 5 apples (Symfony)', ['text' => '{0} There is no apples|{1} There is one apple|]1,Inf] There is %count% apples (%name%)']],