+++ src/Symfony/Component/Translation/DataCollectorTranslator.php
@@ -46,10 +46,12 @@ class DataCollectorTranslator implements TranslatorInterface, TranslatorBagInter
 
     /**
      * {@inheritdoc}
+     *
+     * @param string|null $default A default message if the translation had not been found
      */
-    public function trans(?string $id, array $parameters = [], string $domain = null, string $locale = null)
+    public function trans(?string $id, array $parameters = [], string $domain = null, string $locale = null/*, ?string $default = null*/)
     {
-        $trans = $this->translator->trans($id = (string) $id, $parameters, $domain, $locale);
+        $trans = $this->translator->trans($id = (string) $id, $parameters, $domain, $locale, \func_get_args()[4] ?? null);
         $this->collectMessage($locale, $domain, $id, $trans, $parameters);
 
         return $trans;